requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
psycopg2-binary
Flask-Migrate
reportlab